refactor: :recycle: restructured preprocessing package in preparation of making more robust
