+++ README.md
@@ -11,3 +11,6 @@ TODO: run ^ in a docker container in build time
 Use QGIS to generate datapoints ( possibly on pre split dataset of on land and on water points -> point generation) 
 
 Source: https://www.arcgis.com/home/item.html?id=e750071279bf450cbd510454a80f2e63
+Dataset is then inverted so there are land polygons not water polygons
+
+To do: auto pull and automate the inversion
