+++ preprocessing/testgen/testgen.py
@@ -1,0 +1,10 @@
+from qgis import * 
+
+
+# processing.run("native:randompointsinextent", {'EXTENT':'-130.502597489,-74.202069087,6.177840159,53.427840159 [EPSG:4326]','POINTS_NUMBER':1,'MIN_DISTANCE':0,'TARGET_CRS':QgsCoordinateReferenceSystem('EPSG:4326'),'MAX_ATTEMPTS':200,'OUTPUT':'TEMPORARY_OUTPUT'})
+
+
+def create_test(project: QgsProject):
+    
+    
+    return
